Fix more warnings in MSVC

+++ src/iarray_expression.c
@@ -29,12 +29,12 @@ struct iarray_expression_s {
     iarray_context_t *ctx;
     ina_str_t expr;
     int32_t nchunks;
-    int32_t blocksize;
+    int64_t blocksize;
     int32_t typesize;
     int32_t chunksize;
     int64_t nbytes;
     int nvars;
-    int max_out_len;
+    int64_t max_out_len;
     te_expr *texpr;
     iarray_temporary_t **temp_vars;
     iarray_container_t *out;
@@ -163,7 +163,7 @@ INA_API(ina_rc_t) iarray_expr_compile(iarray_expression_t *e, const char *expr)
                 free(chunk);
             }
             e->chunksize = (int32_t) chunksize;
-            e->blocksize = (int32_t) blocksize;
+            e->blocksize = (int64_t) blocksize;
         }
         else if (e->ctx->cfg->eval_flags == IARRAY_EXPR_EVAL_ITERCHUNK ||
                  e->ctx->cfg->eval_flags == IARRAY_EXPR_EVAL_ITERBLOSC) {
@@ -184,7 +184,7 @@ INA_API(ina_rc_t) iarray_expr_compile(iarray_expression_t *e, const char *expr)
     // TODO: make this more general and accept multidimensional containers
     iarray_dtshape_t dtshape_var = {0};  // initialize to 0s
     dtshape_var.ndim = 1;
-    int temp_var_dim0 = 0;
+    int64_t temp_var_dim0 = 0;
     if (e->ctx->cfg->eval_flags == IARRAY_EXPR_EVAL_ITERBLOCK) {
         temp_var_dim0 = e->blocksize / e->typesize;
     } else if (e->ctx->cfg->eval_flags == IARRAY_EXPR_EVAL_ITERCHUNK ||
@@ -425,7 +425,7 @@ INA_API(ina_rc_t) iarray_eval(iarray_expression_t *e, iarray_container_t *ret)
         // are passed the tinyexpr evaluator.
         // In the future we may want to get rid of the cost of creating/destroying the thread per every chunk.
         // One possibility is to use pthreads, but this would require more complex code, so we need to discuss it more.
-        int32_t blocksize = e->blocksize;
+        int64_t blocksize = e->blocksize;
         int32_t chunksize = e->chunksize;
 
         // Create and initialize an iterator per each variable
@@ -500,7 +500,7 @@ omp_set_num_threads(e->ctx->cfg->max_num_threads);
 //{
 //#endif
             // Do a possible last evaluation with the leftovers
-            int leftover = out_items * e->typesize - nblocks * blocksize;
+            int64_t leftover = out_items * e->typesize - nblocks * blocksize;
             if (leftover > 0) {
                 for (int nvar = 0; nvar < nvars; nvar++) {
                     e->temp_vars[nvar]->data = (char *) iter_value[nvar].block_pointer + nblocks * blocksize;
@@ -641,7 +641,7 @@ static iarray_temporary_t* _iarray_op(iarray_expression_t *expr, iarray_temporar
 
     switch (dtshape.dtype) {
         case IARRAY_DATA_TYPE_DOUBLE: {
-            int len = expr->max_out_len == 0 ? (int)(out->size / sizeof(double)) : expr->max_out_len;
+            int64_t len = expr->max_out_len == 0 ? (int)(out->size / sizeof(double)) : expr->max_out_len;
             if (scalar) {
                 switch(op) {
                 case IARRAY_OPERATION_TYPE_ADD:
@@ -725,7 +725,7 @@ static iarray_temporary_t* _iarray_op(iarray_expression_t *expr, iarray_temporar
         }
         break;
         case IARRAY_DATA_TYPE_FLOAT: {
-            int len = expr->max_out_len == 0 ? (int)(out->size / sizeof(float)) : expr->max_out_len;
+            int64_t len = expr->max_out_len == 0 ? (int)(out->size / sizeof(float)) : expr->max_out_len;
             if (scalar) {
                 switch(op) {
                 case IARRAY_OPERATION_TYPE_ADD:
